refactor: define Epoch type
