+++ packages/automated-dispute/src/types/epochs.ts
@@ -1,0 +1,14 @@
+import { Timestamp } from "@ebo-agent/shared";
+
+/**
+ * Type representing an epoch's data.
+ *
+ * @property {bigint} epoch - epoch number
+ * @property {bigint} epochFirstBlockNumber - number of the first block of the epoch
+ * @property {Timestamp} epochStartTimestamp - timestamp of the first block of the epoch
+ */
+export type Epoch = {
+    epoch: bigint;
+    epochFirstBlockNumber: bigint;
+    epochStartTimestamp: Timestamp;
+};
