+++ packages/automated-dispute/src/types/index.ts
@@ -1,3 +1,4 @@
+export * from "./epochs.js";
 export * from "./events.js";
 export * from "../interfaces/protocolProvider.js";
 export * from "./prophet.js";
